refactor: extract acceptor logic into crate
