+++ crates/ironrdp-acceptor/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ironrdp-server"
+name = "ironrdp-acceptor"
 version = "0.1.0"
 readme = "README.md"
 description = ""
@@ -16,13 +16,9 @@ doctest = false
 test = false
 
 [dependencies]
-anyhow = "1.0"
-bytes = "1"
 tokio = { version = "1", features = ["full"] }
 tokio-rustls = "0.24"
-async-trait = "0.1"
 ironrdp-pdu.workspace = true
+ironrdp-connector.workspace = true
 ironrdp-tokio.workspace = true
-ironrdp-acceptor.workspace = true
-ironrdp-graphics.workspace = true
 tracing.workspace = true
