+++ crates/ironrdp-acceptor/README.md
@@ -1,0 +1,4 @@
+# IronRDP Acceptor
+
+State machine for the server connection acceptance sequence.
+Since ironrdp_server is based on tokio, this crate also includes a tokio driver for the state machine.
