+++ crates/ironrdp-acceptor/src/util.rs
