+++ crates/ironrdp-server/src/lib.rs
@@ -1,7 +1,6 @@
 #[macro_use]
 extern crate tracing;
 
-mod acceptor;
 mod builder;
 mod capabilities;
 mod display;
@@ -12,3 +11,5 @@ mod server;
 pub use display::*;
 pub use handler::*;
 pub use server::*;
+
+pub use ironrdp_acceptor as acceptor;
