+++ crates/ironrdp/Cargo.toml
@@ -22,6 +22,7 @@ connector = ["dep:ironrdp-connector"]
 session = ["dep:ironrdp-session"]
 graphics = ["dep:ironrdp-graphics"]
 input = ["dep:ironrdp-input"]
+server = ["dep:ironrdp-server"]
 
 [dependencies]
 ironrdp-pdu = { workspace = true, optional = true }
@@ -29,3 +30,4 @@ ironrdp-connector = { workspace = true, optional = true }
 ironrdp-session = { workspace = true, optional = true }
 ironrdp-graphics = { workspace = true, optional = true }
 ironrdp-input = { workspace = true, optional = true }
+ironrdp-server = { workspace = true, optional = true }
