+++ crates/ironrdp/src/lib.rs
@@ -10,5 +10,7 @@ pub use ironrdp_graphics as graphics;
 pub use ironrdp_input as input;
 #[cfg(feature = "pdu")]
 pub use ironrdp_pdu as pdu;
+#[cfg(feature = "server")]
+pub use ironrdp_server as server;
 #[cfg(feature = "session")]
 pub use ironrdp_session as session;
